conslidate browser and render lib files
